requirements.txt
pytz~=2020.5
Flask~=2.0.1
discord~=1.7.3
requests~=2.25.1
aiohttp~=3.7.3
bs4~=0.0.1
beautifulsoup4~=4.9.3
psycopg2-binary~=2.9.1
pandas~=1.3.3